+++ app/helpers/pagination.js
@@ -1,0 +1,4 @@
+export {
+  default,
+  pagination,
+} from '@lblod/ember-rdfa-editor-lblod-plugins/helpers/pagination';